Renamed DomTranslator to DomTranslationManager and renamed usages of translationChild appropriately
