+++ extension/ts/content-scripts/dom-translation-content-script.js/DomTranslationManager.spec.ts
